# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 5.0.0
  description: KOWRI V5.0 - Digital Financial Infrastructure for African Markets
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List all transactions
      parameters:
        - name: walletId
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [deposit, transfer, loan_disbursement, loan_repayment, subscription, tontine_contribution, tontine_payout, merchant_payment]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, completed, failed, reversed]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Transactions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"
...
  /tontines:
    get:
      operationId: listTontines
      tags: [tontines]
      summary: List all tontines (group savings)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, completed, pending]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Tontines list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TontineListResponse"
    post:
      operationId: createTontine
      tags: [tontines]
      summary: Create a new tontine group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTontineRequest"
      responses:
        "201":
          description: Tontine created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tontine"
  /tontines/{tontineId}:
    get:
      operationId: getTontine
      tags: [tontines]
      summary: Get tontine by ID
      parameters:
        - name: tontineId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tontine details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TontineDetail"
  /credit/scores:
    get:
      operationId: listCreditScores
      tags: [credit]
      summary: List credit scores
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Credit scores list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditScoreListResponse"
  /credit/scores/{userId}:
    get:
      operationId: getCreditScore
      tags: [credit]
      summary: Get credit score for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Credit score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditScore"
  /credit/loans:
    get:
      operationId: listLoans
      tags: [credit]
      summary: List all loans
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, disbursed, repaid, defaulted]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Loans list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanListResponse"
    post:
      operationId: applyForLoan
      tags: [credit]
      summary: Apply for a micro-loan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanApplicationRequest"
      responses:
        "201":
          description: Loan application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /credit/loans/{loanId}:
    get:
      operationId: getLoan
      tags: [credit]
      summary: Get loan by ID
      parameters:
        - name: loanId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Loan details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /merchants:
    get:
      operationId: listMerchants
      tags: [merchants]
      summary: List all merchants
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, suspended, pending_approval]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Merchants list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantListResponse"
    post:
      operationId: createMerchant
      tags: [merchants]
      summary: Register a new merchant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMerchantRequest"
      responses:
        "201":
          description: Merchant registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Merchant"
...
  /analytics/transactions:
    get:
      operationId: getTransactionAnalytics
      tags: [analytics]
      summary: Get transaction volume analytics
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d]
            default: 30d
      responses:
        "200":
          description: Transaction analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionAnalytics"
  /analytics/ledger:
    get:
      operationId: getLedgerEntries
      tags: [analytics]
      summary: Get ledger entries for audit
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Ledger entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LedgerListResponse"
components:
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required:
        - page
        - limit
        - total
        - totalPages
...
    Transaction:
      type: object
      properties:
        id:
          type: string
        fromWalletId:
          type: string
          nullable: true
        toWalletId:
          type: string
          nullable: true
        amount:
          type: number
        currency:
          type: string
        type:
          type: string
          enum: [deposit, transfer, loan_disbursement, loan_repayment, subscription, tontine_contribution, tontine_payout, merchant_payment]
        status:
          type: string
          enum: [pending, completed, failed, reversed]
        reference:
          type: string
        description:
          type: string
          nullable: true
        metadata:
          type: object
          nullable: true
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - amount
        - currency
        - type
        - status
        - reference
        - createdAt
    TransactionListResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required:
        - transactions
        - pagination
    Tontine:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        contributionAmount:
          type: number
        currency:
          type: string
        frequency:
          type: string
          enum: [weekly, biweekly, monthly]
        memberCount:
          type: integer
        maxMembers:
          type: integer
        currentRound:
          type: integer
        totalRounds:
          type: integer
        status:
          type: string
          enum: [active, completed, pending]
        adminUserId:
          type: string
        nextPayoutDate:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - contributionAmount
        - currency
        - frequency
        - memberCount
        - maxMembers
        - currentRound
        - totalRounds
        - status
        - adminUserId
        - createdAt
    TontineDetail:
      allOf:
        - $ref: "#/components/schemas/Tontine"
        - type: object
          properties:
            members:
              type: array
              items:
                $ref: "#/components/schemas/TontineMember"
            totalContributed:
              type: number
    TontineMember:
      type: object
      properties:
        userId:
          type: string
        userName:
          type: string
        payoutOrder:
          type: integer
        hasReceivedPayout:
          type: boolean
        contributionsCount:
          type: integer
      required:
        - userId
        - userName
        - payoutOrder
        - hasReceivedPayout
        - contributionsCount
    TontineListResponse:
      type: object
      properties:
        tontines:
          type: array
          items:
            $ref: "#/components/schemas/Tontine"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required:
        - tontines
        - pagination
    CreateTontineRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        contributionAmount:
          type: number
        currency:
          type: string
        frequency:
          type: string
          enum: [weekly, biweekly, monthly]
        maxMembers:
          type: integer
        adminUserId:
          type: string
      required:
        - name
        - contributionAmount
        - currency
        - frequency
        - maxMembers
        - adminUserId
    CreditScore:
      type: object
      properties:
        userId:
          type: string
        score:
          type: integer
        tier:
          type: string
          enum: [bronze, silver, gold, platinum]
        maxLoanAmount:
          type: number
        interestRate:
          type: number
        factors:
          type: object
          properties:
            paymentHistory:
              type: integer
            savingsRegularity:
              type: integer
            transactionVolume:
              type: integer
            tontineParticipation:
              type: integer
            networkScore:
              type: integer
        lastUpdated:
          type: string
          format: date-time
      required:
        - userId
        - score
        - tier
        - maxLoanAmount
        - interestRate
        - lastUpdated
    CreditScoreListResponse:
      type: object
      properties:
        scores:
          type: array
          items:
            $ref: "#/components/schemas/CreditScore"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required:
        - scores
        - pagination
    Loan:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        walletId:
          type: string
        amount:
          type: number
        currency:
          type: string
        interestRate:
          type: number
        termDays:
          type: integer
        status:
          type: string
          enum: [pending, approved, disbursed, repaid, defaulted]
        amountRepaid:
          type: number
        dueDate:
          type: string
          format: date-time
          nullable: true
        disbursedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - walletId
        - amount
        - currency
        - interestRate
        - termDays
        - status
        - amountRepaid
        - createdAt
    LoanListResponse:
      type: object
      properties:
        loans:
          type: array
          items:
            $ref: "#/components/schemas/Loan"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required:
        - loans
        - pagination
    LoanApplicationRequest:
      type: object
      properties:
        userId:
          type: string
        walletId:
          type: string
        amount:
          type: number
        currency:
          type: string
        termDays:
          type: integer
        purpose:
          type: string
      required:
        - userId
        - walletId
        - amount
        - currency
        - termDays
        - purpose
    Merchant:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        businessName:
          type: string
        businessType:
          type: string
        status:
          type: string
          enum: [active, suspended, pending_approval]
        walletId:
          type: string
        apiKey:
          type: string
          nullable: true
        country:
          type: string
        totalRevenue:
          type: number
        transactionCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - businessName
        - businessType
        - status
        - walletId
        - country
        - totalRevenue
        - transactionCount
        - createdAt
    MerchantListResponse:
      type: object
      properties:
        merchants:
          type: array
          items:
            $ref: "#/components/schemas/Merchant"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required:
        - merchants
        - pagination
    CreateMerchantRequest:
      type: object
      properties:
        userId:
          type: string
        businessName:
          type: string
        businessType:
          type: string
        country:
          type: string
      required:
        - userId
        - businessName
        - businessType
        - country
...
    TransactionDataPoint:
      type: object
      properties:
        date:
          type: string
        volume:
          type: number
        count:
          type: integer
        type:
          type: string
      required:
        - date
        - volume
        - count
        - type
    TransactionAnalytics:
      type: object
      properties:
        period:
          type: string
        dataPoints:
          type: array
          items:
            $ref: "#/components/schemas/TransactionDataPoint"
        totalVolume:
          type: number
        totalCount:
          type: integer
        byType:
          type: object
          additionalProperties:
            type: number
      required:
        - period
        - dataPoints
        - totalVolume
        - totalCount
        - byType
    LedgerEntry:
      type: object
      properties:
        id:
          type: string
        transactionId:
          type: string
        accountId:
          type: string
        accountType:
          type: string
        debitAmount:
          type: number
        creditAmount:
          type: number
        currency:
          type: string
        eventType:
          type: string
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - transactionId
        - accountId
        - accountType
        - debitAmount
        - creditAmount
        - currency
        - eventType
        - createdAt
    LedgerListResponse:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/LedgerEntry"
        pagination:
          $ref: "#/components/schemas/Pagination"
        totalDebits:
          type: number
        totalCredits:
          type: number
      required:
        - entries
        - pagination
        - totalDebits
        - totalCredits