# lib/api-zod/src/generated/types/creditScore.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KOWRI V5.0 - Digital Financial Infrastructure for African Markets
 * OpenAPI spec version: 5.0.0
 */
import type { CreditScoreFactors } from "./creditScoreFactors";
import type { CreditScoreTier } from "./creditScoreTier";

export interface CreditScore {
  userId: string;
  score: number;
  tier: CreditScoreTier;
  maxLoanAmount: number;
  interestRate: number;
  factors?: CreditScoreFactors;
  lastUpdated: Date;
}